moving forward to working GraphStorage - still delete+clone tests fail

+++ src/main/java/de/jetsli/graph/storage/ListOfArrays.java
@@ -91,7 +91,7 @@ public class ListOfArrays {
         };
     }
 
-    public int capacity() {
+    public long capacity() {
         return refs.capacity() + entries.capacity();
     }
